adding stubs to avoiding duplicated test codes
