+++ tests/test_plugin_driver.c
@@ -64,29 +64,8 @@ void reset_mocks(void)
     mock_free_call_count                = 0;
 }
 
-// Define external buffer variables that plugin_driver.c expects
-// These are normally defined in image_tables.c
-IEC_BOOL *bool_input[BUFFER_SIZE][8];
-IEC_BOOL *bool_output[BUFFER_SIZE][8];
-IEC_BYTE *byte_input[BUFFER_SIZE];
-IEC_BYTE *byte_output[BUFFER_SIZE];
-IEC_UINT *int_input[BUFFER_SIZE];
-IEC_UINT *int_output[BUFFER_SIZE];
-IEC_UDINT *dint_input[BUFFER_SIZE];
-IEC_UDINT *dint_output[BUFFER_SIZE];
-IEC_ULINT *lint_input[BUFFER_SIZE];
-IEC_ULINT *lint_output[BUFFER_SIZE];
-IEC_UINT *int_memory[BUFFER_SIZE];
-IEC_UDINT *dint_memory[BUFFER_SIZE];
-IEC_ULINT *lint_memory[BUFFER_SIZE];
-
-// Mock implementation for plugin_manager_destroy
-// This is normally defined in plcapp_manager.c
-void plugin_manager_destroy(PluginManager *manager)
-{
-    (void)manager; // Suppress unused parameter warning
-    // Mock implementation - do nothing
-}
+// Note: External buffer variables and plugin_manager_destroy are now defined in
+// tests/support/test_plugin_driver_stubs.c and will be linked automatically
 
 void setUp(void)
 {
